requirements..txt
fastapi
uvicorn
python-multipart
numpy
tensorflow
Pillow
streamlit
requests
huggingface_hub